(PC-42798) feat(profile): Rename BonificationIncorrectLink > IncorrectLinkPage
