+++ src/cheatcodes/pages/features/bonification/CheatcodesNavigationBonification.tsx
@@ -98,11 +98,6 @@ const bonificationCheatcodeCategory: CheatcodeCategory = {
     },
     {
       id: uuidv4(),
-      title: 'BonificationIncorrectLink',
-      navigationTarget: getSubscriptionPropConfig('BonificationIncorrectLink'),
-    },
-    {
-      id: uuidv4(),
       title: 'BonificationGranted',
       navigationTarget: { screen: 'BonificationGranted' },
     },
